Split out AggregationStrategy
